Add zip downloading, and redo rollup config again
We needed a lot of plugins in order to get jszip building correctly, so
I decided to mess around with this again. Anyway, we're now doing
chunked outputs a nicer way, which means we don't need all those
explicit files.

Does complicate building a little more, as we're now emitting
'classes.js' into a shared folder along with typescript, but it works
:).