Change CORS proxy
As part of us dropping Cloudflare everywhere. I'd love to use a version
managed by us, but not clear what hosting alternatives - fly.io maybe?

I'm not thrilled about cors.sh (all of these services are a little
transient after all), but it seemed the best of the lot. I considered
corsproxy.io (it does look nice!), but it appears to inject some guff
into the footer of the page which well ... no thanks.

I'll probably re-evaluate this in a week or so. For now, just want
something working!